requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
asyncpg==0.31.0
certifi==2026.2.25
click==8.3.2
exceptiongroup==1.3.1
fastapi==0.136.0
filelock==3.28.0
fsspec==2026.3.0
greenlet==3.4.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==26.1
pillow==12.2.0
pluggy==1.6.0
pydantic==2.13.2
pydantic_core==2.46.2
Pygments==2.20.0
pytest==9.0.3
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
rich==15.0.0
safetensors==0.7.0
shellingham==1.5.4
SQLAlchemy==2.0.49
starlette==1.0.0
sympy==1.14.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
